+++ .setup/index.js
@@ -1,9 +1,11 @@
 'use strict';
 
-var clone = require('./utilities').clone;
+var clone = require('./lib').clone;
 var setPlatform = require('./platform');
 var setEnvironment = require('./environment');
 
+// Set up config by taking the given base configuration and split
+// it into multiple pieces based on overrides.
 module.exports = function(config) {
   var newConfig = {
     overrides: {}