+++ .task/lib/execute.js
@@ -5,6 +5,13 @@ var spawn = require('child_process').spawn;
 var execute = module.exports = Object.create(null);
 
 execute.execute = function(args, options) {
+  // Ensure that stdio io is inherited (allows color preservation)
+  if(options == null) {
+    options = {};
+  }
+  options.stdio = 'inherit';
+
+  // Run spawn and use Q for promises
   var deferred = Q.defer();
   var child = spawn(this.command, args, options);
   child.on('exit', function() {
@@ -13,7 +20,5 @@ execute.execute = function(args, options) {
   child.on('error', function(err) {
     deferred.reject.apply(deferred, arguments);
   });
-  child.stdout.pipe(process.stdout);
-  child.stderr.pipe(process.stderr);
   return deferred.promise;
 };