+++ .task/lib/index.js
@@ -1,0 +1,21 @@
+'use strict';
+
+// Iterate over each task that is declared
+var forEachTask = exports.forEachTask = function(callback, command, tasks) {
+  if(tasks == null) {
+    tasks = require('..');
+  }
+  if(typeof tasks !== 'object') {
+    return;
+  }
+
+  if(tasks.description && tasks.task) {
+    callback(command, tasks.description, tasks.task);
+  }
+  else {
+    for(var index in tasks) {
+      var subCommand = command ? command + ':' : '';
+      forEachTask(callback, subCommand + index, tasks[index]);
+    }
+  }
+};