+++ .task/tasks/bower.js
@@ -5,6 +5,7 @@ var bower = module.exports = Object.create(execute);
 
 bower.command = './node_modules/.bin/bower';
 
+// Just call the execte command and return its promise
 bower.install = function() {
   return this.execute(['install']);
 };